# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  jsonreqresp.BuyTicketRequest:
    properties:
      CustomerEmail:
        example: myname@test.ru
        maxLength: 100
        type: string
      cntTickets:
        example: 1
        minimum: 0
        type: integer
      customerName:
        example: myname
        maxLength: 100
        type: string
      eventID:
        example: b10f841d-ba75-48df-a9cf-c86fc9bd3041
        type: string
    required:
    - cntTickets
    - eventID
    type: object
...
  jsonreqresp.ConfirmCancelTxRequest:
    properties:
      txID:
        example: b10f841d-ba75-48df-a9cf-c86fc9bd3041
        type: string
    required:
    - txID
    type: object
...
  jsonreqresp.TicketPurchaseResponse:
    properties:
      customerEmail:
        type: string
      customerName:
        type: string
      eventId:
        type: string
      id:
        type: string
      purchaseDate:
        type: string
      userId:
        type: string
    type: object
  jsonreqresp.TxTicketPurchaseResponse:
    properties:
      cntTickets:
        type: integer
      expiredAt:
        type: string
      ticketPurchase:
        $ref: '#/definitions/jsonreqresp.TicketPurchaseResponse'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для системы учета произведений искусств
  title: Музей
  version: "1.0"
paths:
...
  /guest/tickets:
    post:
      consumes:
      - application/json
      description: Покупка билетов на указанное мероприятие
      parameters:
      - description: Данные для покупки билетов
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jsonreqresp.BuyTicketRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Данные покупки сохраняются в cookie
          schema:
            $ref: '#/definitions/jsonreqresp.TxTicketPurchaseResponse'
        "400":
          description: Неверный формат запроса
        "401":
          description: Не авторизован
        "404":
          description: Мероприятие не найдено
        "409":
          description: Нет доступных билетов
        "410":
          description: Транзакция просрочена
      summary: Покупка билетов
      tags:
      - Билеты
  /guest/tickets/cancel:
    put:
      consumes:
      - application/json
      description: Отменяет ожидающую транзакцию покупки билетов
      parameters:
      - description: ID транзакции
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jsonreqresp.ConfirmCancelTxRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Покупка отменена
        "400":
          description: Неверный запрос
        "404":
          description: Транзакция не найдена
        "410":
          description: Транзакция просрочена
      summary: Отменить покупку
      tags:
      - Билеты
  /guest/tickets/confirm:
    put:
      consumes:
      - application/json
      description: Подтверждает ожидающую транзакцию покупки билетов
      parameters:
      - description: ID транзакции
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jsonreqresp.ConfirmCancelTxRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Покупка подтверждена
        "400":
          description: Неверный запрос
        "404":
          description: Транзакция не найдена
        "410":
          description: Транзакция просрочена
      summary: Подтвердить покупку
      tags:
      - Билеты
...
swagger: "2.0"